added stock to control book's amount

+++ src/main/java/kz/halykacademy/bookstore/service/impl/OrderServiceImpl.java
@@ -2,25 +2,25 @@ package kz.halykacademy.bookstore.service.impl;
 
 import kz.halykacademy.bookstore.dto.OrderDTO;
 import kz.halykacademy.bookstore.dto.SaveOrderDTO;
-import kz.halykacademy.bookstore.dto.UpdateOrderDTO;
+import kz.halykacademy.bookstore.dto.UpdateOrderByAdminDTO;
 import kz.halykacademy.bookstore.entity.*;
-import kz.halykacademy.bookstore.errors.AuthorizationException;
-import kz.halykacademy.bookstore.errors.OrderInvalidValueException;
+import kz.halykacademy.bookstore.errors.InvalidValueException;
 import kz.halykacademy.bookstore.errors.ResourceNotFoundeException;
 import kz.halykacademy.bookstore.mapper.BookMapper;
 import kz.halykacademy.bookstore.mapper.OrderMapper;
 import kz.halykacademy.bookstore.repository.BookRepository;
 import kz.halykacademy.bookstore.repository.OrderRepository;
+import kz.halykacademy.bookstore.repository.StockRepository;
 import kz.halykacademy.bookstore.repository.UserRepository;
 import kz.halykacademy.bookstore.service.OrderService;
 import lombok.RequiredArgsConstructor;
-import org.aspectj.weaver.ast.Or;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.*;
+import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
@@ -35,6 +35,7 @@ public class OrderServiceImpl implements OrderService {
     private final BookRepository bookRepository;
     private final OrderMapper orderMapper;
     private final BookMapper bookMapper;
+    private final StockRepository stockRepository;
 
 
     @Override
@@ -57,25 +58,12 @@ public class OrderServiceImpl implements OrderService {
         User user = userRepository.findByLogin(userDetails.getUsername()).orElseThrow((Supplier<Throwable>) () ->
                 new ResourceNotFoundeException("user not founded"));
 
-        List<Books> books = bookRepository.getBooksByListId(orderDTO.getBookList());
+        List<Books> books = checkOrder(orderDTO);
+        int sum = checkSum(books, orderDTO);
 
-        if (books.isEmpty()) {
-            throw new OrderInvalidValueException("order cannot be empty");
-        }
-        for (Books book : books) {
-            if (book.isDeleted()) {
-                throw new OrderInvalidValueException("book with id %s deleted".formatted(book.getBookId()));
-            }
-        }
-        List<Integer> sumList = books.stream().map(Books::getPrice).toList();
-
-        int sum = (sumList.stream().mapToInt(a -> a).sum());
-        if (sum > 10000) {
-            throw new OrderInvalidValueException("order exceeded the limit");
-        }
         Order saveOrder = orderRepository.saveAndFlush(
                 new Order(
-                        orderDTO.getOrderId(),
+                        orderDTO.getId(),
                         user,
                         books,
                         sum,
@@ -84,11 +72,10 @@ public class OrderServiceImpl implements OrderService {
                 )
         );
         return orderMapper.toDTO(saveOrder);
-
     }
 
     @Override
-    public OrderDTO updateOrderByAdmin(UpdateOrderDTO orderDTO) throws Throwable {
+    public OrderDTO updateOrderByAdmin(UpdateOrderByAdminDTO orderDTO) throws Throwable {
 
         UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
@@ -101,7 +88,7 @@ public class OrderServiceImpl implements OrderService {
             throw new IllegalArgumentException("you are not allowed to change status of the order");
         }
         if (user.isBlocked()) {
-            throw new OrderInvalidValueException("user with id %s blocked".formatted(orderDTO.getUserId()));
+            throw new InvalidValueException("user with id %s blocked".formatted(orderDTO.getUserId()));
         }
         Order updateOrderStatus = orderRepository.saveAndFlush(
                 new Order(
@@ -118,38 +105,24 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public OrderDTO updateOrderByUser(UpdateOrderDTO orderDTO) throws Throwable {
+    public OrderDTO updateOrderByUser(SaveOrderDTO orderDTO) throws Throwable {
         UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
 
         User user = userRepository.findByLogin(userDetails.getUsername()).orElseThrow((Supplier<Throwable>) () ->
-                new ResourceNotFoundeException("user with id %s not founded".formatted(orderDTO.getUserId())));  //id юзера который вошел
-        Order order = orderRepository.findById(orderDTO.getOrderId()).orElseThrow((Supplier<Throwable>) () ->
-                new ResourceNotFoundeException("order with id %s not founded".formatted(orderDTO.getOrderId()))); // id юзера в заказе
+                new ResourceNotFoundeException("you need to log in"));
+        Order order = orderRepository.findById(orderDTO.getId()).orElseThrow((Supplier<Throwable>) () ->
+                new ResourceNotFoundeException("order with id %s not founded".formatted(orderDTO.getId())));
 
-        if (Objects.equals(user.getUser_id(), order.getUser().getUser_id())) {
+        if (Objects.equals(user.getId(), order.getUser().getId())) {
             if (order.getStatus().name().contains("CREATED")) {
 
-                List<Books> books = bookRepository.getBooksByListId(orderDTO.getBookList());
-
-                if (books.isEmpty()) {
-                    throw new OrderInvalidValueException("order cannot be empty");
-                }
-                for (Books book : books) {
-                    if (book.isDeleted()) {
-                        throw new OrderInvalidValueException("book with id %s deleted".formatted(book.getBookId()));
-                    }
-                }
-                List<Integer> sumList = books.stream().map(Books::getPrice).toList();
-
-                int sum = (sumList.stream().mapToInt(a -> a).sum());
-                if (sum > 10000) {
-                    throw new OrderInvalidValueException("order amount exceeded the limit of 10.000");
-                }
+                List<Books> books = checkOrder(orderDTO);
+                int sum = checkSum(books, orderDTO);
 
                 Order updateOrderList = orderRepository.saveAndFlush(
                         new Order(
-                                orderDTO.getOrderId(),
+                                orderDTO.getId(),
                                 order.getUser(),
                                 books,
                                 sum,
@@ -159,7 +132,7 @@ public class OrderServiceImpl implements OrderService {
                 );
                 return orderMapper.toDTO(updateOrderList);
             } else
-                throw new OrderInvalidValueException("status of the order %s. you are no longer allowed to change the order".formatted(order.getStatus()));
+                throw new InvalidValueException("status of the order %s. you are no longer allowed to change the order".formatted(order.getStatus()));
         } else throw new IllegalArgumentException("its not your order");
 
     }
@@ -167,8 +140,56 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void deleteOrder(Long orderId) throws Throwable {
-        orderRepository.findById(orderId).orElseThrow((Supplier<Throwable>) () ->
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User user = userRepository.findByLogin(userDetails.getUsername()).orElseThrow((Supplier<Throwable>) () ->
+                new ResourceNotFoundeException("you need to log in"));
+        Order order = orderRepository.findById(orderId).orElseThrow((Supplier<Throwable>) () ->
                 new ResourceNotFoundeException("order not founded"));
-        orderRepository.deleteById(orderId);
+        if (Objects.equals(user.getId(), order.getUser().getId())) {
+
+            if (order.getStatus().name().contains("CREATED")) {
+                orderRepository.deleteById(orderId);
+            } else throw new IllegalArgumentException("you can no longer delete the order");
+
+        } else throw new IllegalArgumentException("its not your order");
     }
+
+    private List<Books> checkOrder(SaveOrderDTO orderDTO) {
+        List<Books> books = orderDTO.getBookList().stream().map(bookRepository::getOne).toList();
+
+        if (books.isEmpty()) {
+            throw new InvalidValueException("order cannot be empty");
+        }
+        for (Books book : books) {
+            if (book.isDeleted()) {
+                throw new InvalidValueException("book with id %s deleted".formatted(book.getId()));
+            }
+        }
+        return books;
+    }
+
+    private int checkSum(List<Books> books, SaveOrderDTO orderDTO) {
+        List<Integer> sumList = books.stream().map(Books::getPrice).toList();
+
+        int sum = sumList.stream().mapToInt(a -> a).sum();
+        if (sum > 10000) {
+            throw new InvalidValueException("order exceeded the limit");
+        }
+
+        Map<Long, Long> bookCount = orderDTO.getBookList().stream()
+                .collect(Collectors.groupingBy(Function.identity(), Collectors.counting()));
+
+        List<Long> count = bookCount.values().stream().toList();
+        for (Map.Entry<Long, Long> entry : bookCount.entrySet()) {
+            if (stockRepository.getBookCount(entry.getKey()) - entry.getValue() < 0) {
+                throw new InvalidValueException("all books with id %s send".formatted(entry.getKey()));
+            }
+            stockRepository.updateCount(entry.getValue(), entry.getKey());
+
+        }
+        return sum;
+    }
+
+
 }
